pyproject.toml
[project]
name = "network-idx"
version = "0.1.0"
description = ""
authors = [
    {name = "Eshan Prashar",email = "eprashar@corelogic.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=3.0.1,<4.0.0)",
    "geopandas (>=1.1.2,<2.0.0)",
    "numpy (>=2.4.3,<3.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "shapely (>=2.1.2,<3.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "gcsfs (>=2026.2.0,<2027.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "selenium (>=4.41.0,<5.0.0)",
    "certifi (>=2026.2.25,<2027.0.0)",
    "truststore (>=0.10.4,<0.11.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "google-cloud-storage (>=3.10.0,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "plotly (>=6.7.0,<7.0.0)",
    "pygris (>=0.2.1,<0.3.0)",
    "google-cloud-bigquery (>=3.41.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
